commands: show: Add command to show file from filesystem
The show command will mount a registered filesystem, if need be, and
print the contents of the requested file to STDOUT. An update to
`passphrase:read-passphrase` was made to utilize the pinentry-tty
program so that the show command could be used in command substitution
effectively. The `mount:main` procedure was updated to only call `exit`
in error situations so that it can be used as a library function.