commands/web: Groundwork for `ratel web`
This includes the necessary groundwork for the `ratel web` interface.
The API porition is implemented in this commit which includes listing
all mounts, mounting/unmounting mounts, listing files in a mount and
retrieving individual files from a mount. The fronted for the interface
along with more options regarding SSL will come in later commits.