web/actions:get-mounts: Add computed property and jsify key names
This adds a computed property to the get-mounts action that indicates
whether or not each mount is currently mounted. It also updates the
output to "jsify" the key names so that names with - in them can be
properly addressed with the dot operator in Javascript.

The mounted? procedure was updated to take either a name or an already
retreived configuration value. This is useful to ensure we don't reread
configuration from disk if we already have a value and want to determine
if the mount is mounted.