unshare: Allow execution of commands in an isolated environment
This adds a new flag to the main `ratel` command and a new command to
the `ratel-helper` to aid in the creation of isoloted environments to
run commands in. The isolation is provided by the new command in the
SUID helper that does a few things:

1. Unshares the mount namespace and unmounts all ratel mounts carried
over from the parent namespace.
2. Creates a new session keyring and sets an environment variable to
affect all ratel commands to ensure it is used during mount/umount
operations.
3. execvp's all of the arguments passed to it putting the new process in
this isolated environment

The --unshare option added to the main `ratel` command relaunches the
requested command using the SUID helper making mounted filesystems
"invisible" to the rest of the system.

All of this allows, for instance, running `ratel web` in such a way that
all mount/umount/file actions must go through the web interface and
hides files in mounted filesystems from the rest of the system.