web/frontend: Redesign and include global refresh button
This provides a slight redesign of the UI making it a little wider and
more open. It includes a refresh button that allows refreshing all data
the app might need at any point in time and removes the file specific
refresh button from the file tree view. Code that hints at a global
search bar exists in the Navbar container and will be implemented next
to promote a search-first workflow.