web/frontend: Add global file search to navigation
This implements a file search function that replaces the filter search
available on the individual mount pages and provides a unified search
across all available mounts. This is intended to make it easier/quicker
to get to a specific file in a mount without having to click through to
the mount and find the file in the tree.

A small refactor of the file tree entry controls was made to present the
same controls on the search results.