web/frontend: Add the ability to notify action results
This adds infrastructure for notifying users when things happen in the
app. The first usecase was to notify when an entry was successfully
copied. In the future, other actions should be updated to notify when
they succeed/fail. The CopyFile componenent was made into a container
because it now needs to dispatch PUSH_MESSAGE actions.