web: Improved static asset serving and building
This removes a ton of custom logic in favor of some standard library
functions for matching URL paths to files and simplifies the dispatcher
function that powers the web server. Updates to the Webpack build were
made to accomodate these changes and some minor simplifications were
made to the configuration.